skip other interrupt checks after timer interrupt
If a timer interrupt occurs we must not immediatley check for a io
trap. If we check for io trap intterupt after timer interrupt we
have the possibility of taking a pcb out of the ready q and putting
it back into a waiting queue forever because we never give it cpu
time and so it always has the same pc that always triggers a io
trap.