fixed: find nonempty queue again after starvation pro/demotion
Before this fix the program would sometimes crash because the
we would find the first non empty queque then unluckily the
starvation daemon would be awoken, the daemon would move pcbs
around and then the daemon would finish. Right after the program
would attempt to print the pcb at the head of the queue it had
thought was not empty before the daemon was activated and then
it would crash because the daemon had emptied the queue.